Implement middleware framework
The middleware framework forms the basis of the AlpineBits server.
Its concepts are inspired by koajs.

The middleware consits of:

- middleware: a middleware implements a certain aspect or funcionality.
It is usually invoked by a chain, that contains many middlewares.
A middleware is responsible of calling the next middleware in the
chain, but it doesn't have to.
- chain: the chain invokes the middlewares
- context: the middlewares share a context. The context provides
base information (e.g. about the container) and can be used to
read/write data
- container: the container glues the other parts together and
should e.g. be used to start the middleware chain